Rename field `orientation` to `pageIndicatorOrientation`
