+++ library/src/main/java/com/tapptitude/pageindicator/util/DpValue.kt
@@ -7,4 +7,4 @@ import androidx.annotation.Dimension.DP
  * Marks a value to be interpreted and used as density independent pixels
  */
 @Dimension(unit = DP)
-annotation class DpValue
+internal annotation class DpValue
